# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LoopSpace API - Beta feedback collection platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: List all pages in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
    post:
      operationId: createPage
      tags: [pages]
      summary: Create a new page in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageInput"
      responses:
        "201":
          description: Created page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
  /projects/{projectId}/pages/{pageId}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get a page by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: pageId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Not found
    patch:
      operationId: updatePage
      tags: [pages]
      summary: Update a page
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: pageId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageUpdate"
      responses:
        "200":
          description: Updated page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Not found
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete a page
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: pageId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Page:
      type: object
      required: [id, projectId, name, iframeToken, createdAt, updatedAt]
      properties:
        id:
          type: number
        projectId:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        iframeToken:
          type: string
        iframeUrl:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        feedbackCount:
          type: number
        openCount:
          type: number
    PageInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    PageUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
...