# .migration-backup/lib/api-zod/src/generated/types/page.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LoopSpace API - Beta feedback collection platform
 * OpenAPI spec version: 0.1.0
 */

export interface Page {
  id: number;
  projectId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  iframeToken: string;
  iframeUrl?: string;
  createdAt: string;
  updatedAt: string;
  feedbackCount?: number;
  openCount?: number;
}