# .migration-backup/lib/api-zod/src/generated/types/pageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LoopSpace API - Beta feedback collection platform
 * OpenAPI spec version: 0.1.0
 */

export interface PageInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}